requirements.txt
Flask==3.0.0
numpy==1.26.0
requests==2.31.0
google-generativeai==0.3.1
scikit-learn==1.3.0
gunicorn